fetch real data to the register view
