+++ packages/client/src/v2-events/features/events/fixtures.ts
@@ -143,6 +143,13 @@ export const tennisClubMembershipEvent = {
               description: 'This is the first version of the form'
             }
           },
+          review: {
+            title: {
+              id: 'event.tennis-club-membership.action.declare.form.review.title',
+              defaultMessage: 'Member declaration for {firstname} {surname}',
+              description: 'Title of the form to show in review page'
+            }
+          },
           pages: [
             {
               id: 'applicant',
