+++ packages/commons/src/events/index.ts
@@ -21,3 +21,4 @@ export * from './ActionInput'
 export * from './ActionDocument'
 export * from './EventIndex'
 export * from './TranslationConfig'
+export * from './state'
