+++ packages/events/src/service/indexing/indexing.ts
@@ -12,73 +12,19 @@
 const EVENTS_INDEX = 'events'
 
 import {
-  ActionDocument,
-  CreatedAction,
   EventDocument,
   EventIndex,
-  EventStatus
+  getCurrentEventState
 } from '@opencrvs/commons/events'
 
+import { type estypes } from '@elastic/elasticsearch'
 import { getClient } from '@events/storage'
 import { getOrCreateClient } from '@events/storage/elasticsearch'
-import { type estypes } from '@elastic/elasticsearch'
 import { Transform } from 'stream'
 import { z } from 'zod'
 
-function getStatusFromActions(actions: Array<ActionDocument>) {
-  return actions.reduce<EventStatus>((status, action) => {
-    if (action.type === 'CREATE') {
-      return 'CREATED'
-    }
-    if (action.type === 'DECLARE') {
-      return 'DECLARED'
-    }
-    if (action.type === 'DRAFT') {
-      return 'DRAFT'
-    }
-    return status
-  }, 'CREATED')
-}
-
-function getAssignedUserFromActions(actions: Array<ActionDocument>) {
-  return actions.reduce<null | string>((status, action) => {
-    if (action.type === 'ASSIGN') {
-      return action.assignedTo
-    }
-    if (action.type === 'UNASSIGN') {
-      return null
-    }
-    return status
-  }, null)
-}
-
-function getData(actions: Array<ActionDocument>) {
-  return actions.reduce((status, action) => {
-    return {
-      ...status,
-      ...action.data
-    }
-  }, {})
-}
-
 function eventToEventIndex(event: EventDocument): EventIndex {
-  const creationAction = event.actions.find(
-    (action) => action.type === 'CREATE'
-  ) as CreatedAction
-  const latestAction = event.actions[event.actions.length - 1]
-
-  return {
-    id: event.id,
-    type: event.type,
-    status: getStatusFromActions(event.actions),
-    createdAt: event.createdAt,
-    createdBy: creationAction.createdBy,
-    createdAtLocation: creationAction.createdAtLocation,
-    modifiedAt: latestAction.createdAt,
-    assignedTo: getAssignedUserFromActions(event.actions),
-    updatedBy: latestAction.createdBy,
-    data: getData(event.actions)
-  }
+  return getCurrentEventState(event)
 }
 
 /*
